More splitting into dedicated tests for better isolation
