+++ src/ServiceControl.AcceptanceTests/Audit/When_message_successfully_processed.cs
@@ -1,0 +1,79 @@
+﻿namespace ServiceBus.Management.AcceptanceTests
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using NServiceBus;
+    using NServiceBus.AcceptanceTesting;
+    using NServiceBus.Settings;
+    using NUnit.Framework;
+    using ServiceBus.Management.AcceptanceTests.Contexts;
+    using ServiceControl.CompositeViews.Endpoints;
+
+    public class When_message_successfully_processed : AcceptanceTest
+    {
+        [Test]
+        public void Should_list_the_endpoint_in_the_list_of_known_endpoints()
+        {
+            var context = new MyContext();
+
+            List<EndpointsView> knownEndpoints = null;
+
+            Define(context)
+                .WithEndpoint<Sender>(b => b.Given((bus, c) =>
+                {
+                    bus.Send(new MyMessage());
+                }))
+                .WithEndpoint<Receiver>()
+                .Done(c => TryGetMany("/api/endpoints", out knownEndpoints, m => m.Name == context.EndpointNameOfReceivingEndpoint))
+                .Run(TimeSpan.FromSeconds(20));
+
+            Assert.AreEqual(context.EndpointNameOfReceivingEndpoint, knownEndpoints.Single(e => e.Name == context.EndpointNameOfReceivingEndpoint).Name);
+            Assert.AreEqual(Environment.MachineName, knownEndpoints.Single(e => e.Name == context.EndpointNameOfReceivingEndpoint).HostDisplayName);
+        }
+
+        public class Sender : EndpointConfigurationBuilder
+        {
+            public Sender()
+            {
+                EndpointSetup<DefaultServerWithoutAudit>()
+                    .AddMapping<MyMessage>(typeof(Receiver));
+            }
+        }
+
+        public class Receiver : EndpointConfigurationBuilder
+        {
+            public Receiver()
+            {
+                EndpointSetup<DefaultServerWithAudit>();
+            }
+
+            public class MyMessageHandler : IHandleMessages<MyMessage>
+            {
+                public MyContext Context { get; set; }
+
+                public IBus Bus { get; set; }
+
+                public ReadOnlySettings Settings { get; set; }
+
+                public void Handle(MyMessage message)
+                {
+                    Context.EndpointNameOfReceivingEndpoint = Settings.EndpointName();
+                    Context.MessageId = Bus.CurrentMessageContext.Id;
+                }
+            }
+        }
+
+        [Serializable]
+        public class MyMessage : ICommand
+        {
+        }
+
+        public class MyContext : ScenarioContext
+        {
+            public string MessageId { get; set; }
+
+            public string EndpointNameOfReceivingEndpoint { get; set; }
+        }
+    }
+}
