+++ src/ServiceControl.AcceptanceTests/Audit/When_processed_message_is_imported.cs
@@ -1,0 +1,129 @@
+﻿namespace ServiceBus.Management.AcceptanceTests
+{
+    using System;
+    using System.Linq;
+    using System.Text;
+    using Contexts;
+    using Newtonsoft.Json;
+    using NServiceBus;
+    using NServiceBus.AcceptanceTesting;
+    using NServiceBus.Settings;
+    using NUnit.Framework;
+    using ServiceControl.CompositeViews.Messages;
+    using ServiceControl.Contracts.Operations;
+
+    public class When_processed_message_is_imported : AcceptanceTest
+    {
+        [Test]
+        public void Should_be_accessible_via_the_rest_api()
+        {
+            const string Payload = "PAYLOAD";
+            var context = new MyContext();
+            MessagesView auditedMessage = null;
+            byte[] body = null;
+
+            Define(context)
+                .WithEndpoint<Sender>(b => b.Given((bus, c) =>
+                {
+                    bus.Send(new MyMessage
+                    {
+                        Payload = Payload
+                    });
+                }))
+                .WithEndpoint<Receiver>()
+                .Done(c =>
+                {
+                    if (c.MessageId == null)
+                    {
+                        return false;
+                    }
+                    if (!TryGetSingle("/api/messages?include_system_messages=false&sort=id", out auditedMessage, m => m.MessageId == c.MessageId))
+                    {
+                        return false;
+                    }
+
+                    body = DownloadData(auditedMessage.BodyUrl);
+
+                    return true;
+
+                })
+                .Run(TimeSpan.FromSeconds(40));
+
+            Console.WriteLine(JsonConvert.SerializeObject(auditedMessage));
+
+            Assert.AreEqual(context.MessageId, auditedMessage.MessageId);
+            Assert.AreEqual(MessageStatus.Successful, auditedMessage.Status);
+            Assert.AreEqual(context.EndpointNameOfReceivingEndpoint, auditedMessage.ReceivingEndpoint.Name,
+                "Receiving endpoint name should be parsed correctly");
+
+            Assert.AreNotEqual(Guid.Empty, auditedMessage.ReceivingEndpoint.HostId, "Host id should be set");
+            Assert.False(string.IsNullOrEmpty(auditedMessage.ReceivingEndpoint.Host), "Host display name should be set");
+
+            Assert.AreEqual(typeof(MyMessage).FullName, auditedMessage.MessageType,
+                "AuditMessage type should be set to the FullName of the message type");
+            Assert.False(auditedMessage.IsSystemMessage, "AuditMessage should not be marked as a system message");
+
+            Assert.NotNull(auditedMessage.ConversationId);
+
+            Assert.AreNotEqual(DateTime.MinValue, auditedMessage.TimeSent, "Time sent should be correctly set");
+            Assert.AreNotEqual(DateTime.MinValue, auditedMessage.ProcessedAt, "Processed At should be correctly set");
+
+            Assert.Less(TimeSpan.Zero, auditedMessage.ProcessingTime, "Processing time should be calculated");
+            Assert.Less(TimeSpan.Zero, auditedMessage.CriticalTime, "Critical time should be calculated");
+            Assert.AreEqual(MessageIntentEnum.Send, auditedMessage.MessageIntent, "Message intent should be set");
+
+            var bodyAsString = Encoding.UTF8.GetString(body);
+
+            Assert.True(bodyAsString.Contains(Payload), bodyAsString);
+
+            Assert.AreEqual(body.Length, auditedMessage.BodySize);
+
+            Assert.True(auditedMessage.Headers.Any(_=>_.Key == Headers.MessageId));
+        }
+
+        public class Sender : EndpointConfigurationBuilder
+        {
+            public Sender()
+            {
+                EndpointSetup<DefaultServerWithoutAudit>()
+                    .AddMapping<MyMessage>(typeof(Receiver));
+            }
+        }
+
+        public class Receiver : EndpointConfigurationBuilder
+        {
+            public Receiver()
+            {
+                EndpointSetup<DefaultServerWithAudit>();
+            }
+
+            public class MyMessageHandler : IHandleMessages<MyMessage>
+            {
+                public MyContext Context { get; set; }
+
+                public IBus Bus { get; set; }
+
+                public ReadOnlySettings Settings { get; set; }
+
+                public void Handle(MyMessage message)
+                {
+                    Context.EndpointNameOfReceivingEndpoint = Settings.EndpointName();
+                    Context.MessageId = Bus.CurrentMessageContext.Id;
+                }
+            }
+        }
+
+        [Serializable]
+        public class MyMessage : ICommand
+        {
+            public string Payload { get; set; }
+        }
+
+        public class MyContext : ScenarioContext
+        {
+            public string MessageId { get; set; }
+
+            public string EndpointNameOfReceivingEndpoint { get; set; }
+        }
+    }
+}
